Check if winget exists

+++ setup.ps1
@@ -42,46 +42,50 @@ if (!(New-Object Security.Principal.WindowsPrincipal([Security.Principal.Windows
 # Apparently progress bars drastically slow WebRequest downloads
 $ProgressPreference = 'SilentlyContinue'
 
-# 
-# Dependencies
-#
-# Source: https://github.com/microsoft/winget-cli/issues/1861
+# Check if winget is already installed (possible for newer versions of windows)
 
-# Get 'Desktop Installer'
-Write-Output "Installing dependency 'Microsoft Desktop Installer'"
-Add-AppxPackage -Path https://aka.ms/Microsoft.VCLibs.x64.14.00.Desktop.appx
-
-# Get Microsoft.UI.Xaml
-# Note: nuget is not owned by Microsoft, but is funded by microsoft.:shrug:
-Write-Output "Installing dependency 'Microsoft UI Xaml 2.7'"
-Invoke-WebRequest -Uri https://www.nuget.org/api/v2/package/Microsoft.UI.Xaml/2.7.3 -OutFile $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3.zip'
-Expand-Archive -Path $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3.zip' -DestinationPath $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3'
-Add-AppxPackage $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3\tools\AppX\x64\Release\Microsoft.UI.Xaml.2.7.appx'
-
-Remove-Item $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3.zip'
-Remove-Item $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3' -Recurse
-
-#
-# Winget
-#
-# Source: https://stackoverflow.com/questions/74166150/install-winget-by-the-command-line-powershell
-
-# get latest download url
-Write-Output "Installing winget"
-$URL = "https://api.github.com/repos/microsoft/winget-cli/releases/latest"
-$URL = (Invoke-WebRequest -UseBasicParsing -Uri $URL).Content | ConvertFrom-Json |
-        Select-Object -ExpandProperty "assets" |
-        Where-Object "browser_download_url" -Match '.msixbundle' |
-        Select-Object -ExpandProperty "browser_download_url"
-
-# download
-Invoke-WebRequest -Uri $URL -OutFile $env:USERPROFILE'\Downloads\Setup.msix' -UseBasicParsing
-
-# install
-Add-AppxPackage -Path $env:USERPROFILE'\Downloads\Setup.msix'
-
-# delete file
-Remove-Item $env:USERPROFILE'\Downloads\Setup.msix'
+if(Test-CommandExists "winget") {
+	# 
+	# Dependencies
+	#
+	# Source: https://github.com/microsoft/winget-cli/issues/1861
+	
+	# Get 'Desktop Installer'
+	Write-Output "Installing dependency 'Microsoft Desktop Installer'"
+	Add-AppxPackage -Path https://aka.ms/Microsoft.VCLibs.x64.14.00.Desktop.appx
+	
+	# Get Microsoft.UI.Xaml
+	# Note: nuget is not owned by Microsoft, but is funded by microsoft.:shrug:
+	Write-Output "Installing dependency 'Microsoft UI Xaml 2.7'"
+	Invoke-WebRequest -Uri https://www.nuget.org/api/v2/package/Microsoft.UI.Xaml/2.7.3 -OutFile $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3.zip'
+	Expand-Archive -Path $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3.zip' -DestinationPath $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3'
+	Add-AppxPackage $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3\tools\AppX\x64\Release\Microsoft.UI.Xaml.2.7.appx'
+	
+	Remove-Item $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3.zip'
+	Remove-Item $env:USERPROFILE'\Downloads\microsoft.ui.xaml.2.7.3' -Recurse
+	
+	#
+	# Winget
+	#
+	# Source: https://stackoverflow.com/questions/74166150/install-winget-by-the-command-line-powershell
+	
+	# get latest download url
+	Write-Output "Installing winget"
+	$URL = "https://api.github.com/repos/microsoft/winget-cli/releases/latest"
+	$URL = (Invoke-WebRequest -UseBasicParsing -Uri $URL).Content | ConvertFrom-Json |
+	        Select-Object -ExpandProperty "assets" |
+	        Where-Object "browser_download_url" -Match '.msixbundle' |
+	        Select-Object -ExpandProperty "browser_download_url"
+	
+	# download
+	Invoke-WebRequest -Uri $URL -OutFile $env:USERPROFILE'\Downloads\Setup.msix' -UseBasicParsing
+	
+	# install
+	Add-AppxPackage -Path $env:USERPROFILE'\Downloads\Setup.msix'
+	
+	# delete file
+	Remove-Item $env:USERPROFILE'\Downloads\Setup.msix'
+}
 
 Update-Environment-Path
